--- src/test-config/mocks/FormQuestion.mock.ts
@@ -1,8 +1,0 @@
-import { FormQuestion } from '@/payload-types'
-
-export const formQuestionMock: FormQuestion = {
-  id: '67ff38a56a35e1b6cf43a681',
-  question: 'What do you want out of this project?',
-  updatedAt: new Date().toISOString(),
-  createdAt: new Date().toISOString(),
-}
